+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.8.2] - 2024-01-26
+
+- Feat: Add endpoint to force close ln-dlc channels
+
 ## [1.8.1] - 2024-01-26
 
 - Chore: Enforce minimum quantity to 500 on dlc channel open
@@ -270,7 +274,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Self-Custodial CFD Trading based on DLC and lightning
 
-[Unreleased]: https://github.com/get10101/10101/compare/1.8.1...HEAD
+[Unreleased]: https://github.com/get10101/10101/compare/1.8.2...HEAD
+[1.8.2]: https://github.com/get10101/10101/compare/1.8.1...1.8.2
 [1.8.1]: https://github.com/get10101/10101/compare/1.8.0...1.8.1
 [1.8.0]: https://github.com/get10101/10101/compare/1.7.4...1.8.0
 [1.7.4]: https://github.com/get10101/10101/compare/1.7.3...1.7.4